[SCB-1407] handler-router decoupled spi
