+++ handlers/handler-router/src/main/java/org/apache/servicecomb/router/custom/MicroserviceCache.java
@@ -14,23 +14,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.servicecomb.router.match;
+package org.apache.servicecomb.router.custom;
 
+import java.util.HashMap;
 import java.util.Map;
 
-/**
- * @Author GuoYl123
- * @Date 2019/10/17
- **/
-public interface RouterHeaderFilterExt {
+import org.apache.servicecomb.serviceregistry.RegistryUtils;
+import org.apache.servicecomb.serviceregistry.api.registry.Microservice;
 
-  default int getOrder() {
-    return 0;
+public final class MicroserviceCache {
+
+  private static MicroserviceCache instance = new MicroserviceCache();
+  private Map<String, Microservice> services = new HashMap<>();
+
+  private MicroserviceCache() {
   }
 
-  default boolean enabled() {
-    return true;
+  public static MicroserviceCache getInstance() {
+    return instance;
   }
 
-  Map<String, String> doFilter(Map<String, String> invokeHeader);
+  public Microservice getService(String serviceId) {
+    return services.computeIfAbsent(serviceId, (k) ->
+        RegistryUtils.getMicroservice(serviceId)
+    );
+  }
 }
